Pythonagents/fastapi-mcp-agent/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
httpx
sqlalchemy
databases
asyncpg
python-multipart
pytest
pytest-asyncio
websockets
aiofiles
python-dotenv

# Database
motor
pymongo
beanie

# MCP SDK and adapters
mcp
mcp-client
mcp-server
langchain-mcp-adapters

# LangChain for agent graphs (latest compatible versions)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=3.0.0
langchain-groq>=0.1.0
langgraph>=0.1.0

# Model providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.8.0

# Additional utilities
pathlib2
typing-extensions
asyncio
 